feat: http layer and refine on store module

+++ src/modules/stores/application/use-cases/delete-store.use-case.ts
@@ -1,6 +1,7 @@
 import { Injectable } from "@nestjs/common";
 import { StoreRepository } from "../../domain/repositories/store.repositories";
 import { StoreNotFoundError } from "../errors/store-not-found.error";
+import { StoreHasStockError } from "../errors/store-has-stock.error";
 
 
 @Injectable()
@@ -13,6 +14,11 @@ export class DeleteStoreUseCase {
 
         if (!store) {
             throw new StoreNotFoundError();
+        }
+
+        const hasStock = await this.storeRepository.hasStock(storeId);
+        if (hasStock) {
+            throw new StoreHasStockError();
         }
 
         await this.storeRepository.delete(storeId);
